Refactor: further split query.js
