+++ packages/sync-server/app/hl7fhir/materialised/search/order.js
@@ -2,8 +2,28 @@ import { Sequelize } from 'sequelize';
 import { FHIR_SEARCH_PARAMETERS, FHIR_SEARCH_TOKEN_TYPES } from 'shared/constants';
 import { Unsupported } from 'shared/utils/fhir';
 
+import { findField } from './common';
+
+export function generateOrderClause(query, parameters, FhirResource) {
+  let ordering = [];
+
+  for (const { order, by } of query.get('_sort').flatMap(v => v.value)) {
+    const def = parameters.get(by);
+    if (def.path.length === 0) continue;
+
+    const alternates = def.path.map(([field, ...path]) => {
+      const resolvedPath = [findField(FhirResource, field).field, ...path];
+      return singleOrder(resolvedPath, order, def, FhirResource);
+    });
+
+    ordering = ordering.concat(alternates);
+  }
+
+  return ordering;
+}
+
 // eslint-disable-next-line no-unused-vars
-export function singleOrder(path, order, def, _Model) {
+function singleOrder(path, order, def, _Model) {
   const entirePath = path;
   if (
     def.type === FHIR_SEARCH_PARAMETERS.TOKEN &&
