+++ packages/sync-server/app/hl7fhir/materialised/search/query.js
@@ -1,17 +1,11 @@
 import { last } from 'lodash';
-import { Op } from 'sequelize';
 
 import { MAX_RESOURCES_PER_PAGE } from 'shared/constants';
-import { RESULT_PARAMETER_NAMES } from 'shared/utils/fhir';
 
-import { findField, singleMatch } from './match';
-import { singleOrder } from './order';
-
-export function pushToQuery(query, param, value) {
-  const insert = query.get(param) ?? [];
-  insert.push(value);
-  query.set(param, insert);
-}
+import { pushToQuery } from './common';
+import { generateIncludeClause } from './include';
+import { generateWhereClause } from './where';
+import { generateOrderClause } from './order';
 
 /**
  * @param {*} query The request query Map (normalised)
@@ -24,19 +18,7 @@ export function buildSearchQuery(query, parameters, FhirResource) {
   };
 
   if (query.has('_sort')) {
-    let ordering = [];
-    for (const { order, by } of query.get('_sort').flatMap(v => v.value)) {
-      const def = parameters.get(by);
-      if (def.path.length === 0) continue;
-
-      const alternates = def.path.map(([field, ...path]) => {
-        const resolvedPath = [findField(FhirResource, field).field, ...path];
-        return singleOrder(resolvedPath, order, def, FhirResource);
-      });
-
-      ordering = ordering.concat(alternates);
-    }
-    sql.order = ordering;
+    sql.order = generateOrderClause(query, parameters, FhirResource);
   }
 
   if (query.has('_count')) {
@@ -55,23 +37,7 @@ export function buildSearchQuery(query, parameters, FhirResource) {
 
   // TODO: support _summary and _elements
 
-  const andWhere = [];
-  for (const [name, paramQueries] of query.entries()) {
-    if (RESULT_PARAMETER_NAMES.includes(name)) continue;
-
-    const def = parameters.get(name);
-    if (def.path.length === 0) continue;
-
-    for (const paramQuery of paramQueries) {
-      const alternates = def.path.flatMap(([field, ...path]) => {
-        const resolvedPath = [findField(FhirResource, field).field, ...path];
-        return singleMatch(resolvedPath, paramQuery, def, FhirResource);
-      });
-
-      andWhere.push({ [Op.or]: alternates });
-    }
-  }
-  sql.where = { [Op.and]: andWhere };
+  sql.where = generateWhereClause(query, parameters, FhirResource);
 
   return sql;
 }
